refactor platform agnostic implementation
* add group action loader to task_executor
* add ANSIBLE_ACTION_HANDLER attribute to net_system
* remove plugins/action/net_system
* reorganize plugins/action/network_base

Still more refactoring to do once the connection plugin work is complete
